Don't install puppetlabs repo when MANAGE_REPOS=false
Currently, p-o-i allways tries to install puppet package
from puppetlabs repo. In some cases, as RDO-CI we use p-o-i
to test puppet modules from RDO distro and it's recommended
to test it using puppet version provided in RDO repos which
uses different paths.

This patch provides a way to override this behavior by using
env variables as:

MANAGE_REPOS=false
PUPPET_PKG=puppet
PUPPETFILE_DIR=/etc/puppet/modules

(cherry picked from commit 2887c1652f955db8cf474a4f3d20223f1c790cae)

+++ install_modules.sh
@@ -13,10 +13,12 @@ fi
 if [ "${PUPPET_MAJ_VERSION}" = 3 ]; then
   export PUPPET_BASE_PATH=/etc/puppet
 else
-  export PUPPET_BASE_PATH=/etc/puppetlabs/code
-  export PATH=${PATH}:/opt/puppetlabs/bin
   # Workaround to deploy puppet for beaker jobs
-  sudo -E ln -sfn /opt/puppetlabs/bin/puppet /usr/sbin/puppet
+  if [ -e /opt/puppetlabs/bin/puppet ]; then
+    export PUPPET_BASE_PATH=/etc/puppetlabs/code
+    export PATH=${PATH}:/opt/puppetlabs/bin:/opt/puppetlabs/puppet/bin
+    sudo -E ln -sfn /opt/puppetlabs/bin/puppet /usr/sbin/puppet
+  fi
 fi
 
 export SCRIPT_DIR=$(cd `dirname $0` && pwd -P)
